Separate NewExpense into more components, update field errors
