+++ src/common/_buttons.scss
@@ -44,6 +44,7 @@
   }
 
   &.button--medium {
+    min-height: 36px;
     padding: 8px 12px;
     font-size: 0.875rem;
     font-weight: 500;
