+++ src/common/_form.scss
@@ -87,14 +87,21 @@
   }
 }
 
+.field-error {
+  border: 2px solid v.$red !important;
+}
+
 .field-error-text {
-  display: inline-block;
-  align-self: flex-start;
+  position: absolute;
+  z-index: 4;
+  bottom: -60%;
+  left: 0;
+  right: 0;
+  width: fit-content;
   border: 1px solid v.$red;
   background: rgba(220, 85, 55, 0.05);
   border-radius: 0.25rem;
   padding: 0.5rem;
-  margin: 0 auto;
 
   &.form-validation {
     align-self: center;
