+++ src/components/Input/AmountInput/AmountInput.scss
@@ -1,8 +1,7 @@
 @use '../../../utilities/colors' as v;
 
 .amount-container {
-  overflow: hidden;
-
+  position: relative;
   .amount-input-container {
     display: flex;
     justify-content: center;
@@ -14,7 +13,6 @@
     font-weight: 500;
     font-size: 2.5rem;
   }
-
   .amount-input {
     padding: 0;
     min-height: auto;
@@ -24,9 +22,5 @@
     letter-spacing: inherit;
     text-align: center;
     border: none;
-
-    &.field-error {
-      border: 2px solid v.$red;
-    }
   }
 }
