+++ src/components/Input/TextInput/TextInput.scss
@@ -1,6 +1,7 @@
 @use '../../../utilities/colors' as v;
 
 .text-input-container {
+  position: relative;
   display: flex;
   flex-direction: column;
   gap: 0.75rem;
@@ -13,7 +14,6 @@
 
 .text-input {
   min-height: 38px;
-
   &::placeholder {
     color: #bfbfbf;
   }
