+++ src/pages/NewExpense/NewExpense.js
@@ -1,23 +1,17 @@
 import './NewExpense.scss';
-import { useState, useEffect, useRef } from 'react';
+import { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useAddNewExpenseMutation } from '../../slices/expenseSlice';
 import { useAddNewDebtMutation } from '../../slices/debtSlice';
 import { getAllGroupsUsers } from '../../services/groups';
 import { useAddActivityMutation } from '../../slices/activityApi';
-import Header from '../../components/Header/Header';
-import TextInput from '../../components/Input/TextInput/TextInput';
-import AmountInput from '../../components/Input/AmountInput/AmountInput';
-import RadioSelect from '../../components/Input/RadioSelect/RadioSelect';
-import UserButton from '../../components/User/UserButton/UserButton';
-import Modal from '../../components/Modal/Modal';
-import SelectPeople from '../../components/SelectPeople/SelectPeople';
 import { useNavigate } from 'react-router-dom';
 import { useGetAccountQuery } from '../../slices/accountSlice';
-import AddUsers from './AddUsers/AddUsers';
-import Atom from '../../components/Atom/Atom';
-import DropdownSelect from '../../components/Input/DropdownSelect/DropdownSelect';
-import { CATEGORIES } from '../../constants/categories';
+import Header from '../../components/Header/Header';
+import Modal from '../../components/Modal/Modal';
+import SelectPeople from '../../components/SelectPeople/SelectPeople';
+import NewExpenseDetails from './NewExpenseDetails/NewExpenseDetails';
+import NewExpenseSplit from './NewExpenseSplit/NewExpenseSplit';
 
 const NewExpense = () => {
   const user = useSelector(state => state.auth.user);
@@ -32,22 +26,15 @@ const NewExpense = () => {
   const [splitWithGroup, setSplitWithGroup] = useState(null);
   const [paidBy, setPaidBy] = useState(user?.id);
   const [split, setSplit] = useState('');
+  const [page, setPage] = useState(1);
   const [fieldErrors, setFieldErrors] = useState({});
-  const [category, setCategory] = useState('category');
-  const [categoryImg, setCategoryImg] = useState(null);
   const [openSelectPeople, setOpenSelectPeople] = useState(false);
-  const navigate = useNavigate();
-
-  let categoriesList = useRef(Object.keys(CATEGORIES));
-
-  useEffect(() => {
-    const getCategoryImage = CATEGORIES[category]?.image;
-    setCategoryImg(getCategoryImage);
-  }, [category])
 
   const [addNewExpense, { isExpenseLoading }] = useAddNewExpenseMutation();
   const [addNewDebt, { isDebtLoading }] = useAddNewDebtMutation();
   const [addActivity, { isActivityLoading }] = useAddActivityMutation();
+
+  const navigate = useNavigate();
 
   const handleSubmit = async e => {
     e.preventDefault();
@@ -133,11 +120,6 @@ const NewExpense = () => {
     return formValid;
   };
 
-  const handleAmount = value => {
-    const validInput = value.match(/^(\d*\.{0,1}\d{0,2}$)/);
-    if (validInput) setAmount(value);
-  };
-
   const handleAdd = selected => {
     if (selected?.group_name) {
       setSplitWithGroup(selected);
@@ -160,137 +142,36 @@ const NewExpense = () => {
 
   return (
     <>
-      <Header type="title" title="New expense" classes="transparent" />
+      <Header type="title" title="New expense" classes="transparent" headerLeftFn={
+        page === 1 ? () => navigate(-1) : () => setPage(1)
+      } />
       <div className="expense-container">
         <form className="expense-form" onSubmit={handleSubmit}>
-          <div className="expense-details-page">
-            <div className="expense-input-container">
-              {fieldErrors.amount && (
-                <span className="field-error-text">{fieldErrors.amount}</span>
-              )}
-              <TextInput
-                className="expense-input big"
-                name="description"
-                label="Expense name"
-                value={description}
-                size={description.length > 0 ? description.length + 1 : 13}
-                placeholder="Expense name"
-                fieldError={fieldErrors.description}
-                onFocus={() =>
-                  setFieldErrors({ ...fieldErrors, description: null })
-                }
-                onChange={e => setDescription(e.target.value)}
-              />
-            </div>
-            <Atom 
-              size={120}
-              numOrbitals={4}
-              fade={true}
-              image={
-                <img src={categoryImg ? require(`../../${categoryImg}`) : null} alt={`${category} icon`} style={{
-                  filter: 'brightness(0) invert(1)',
-                }} height={48} width={41} />
-              }
-            >
-              <DropdownSelect
-                options={categoriesList.current}
-                value={category}
-                name="category-select"
-                onChange={(e) => setCategory(e.target.value)}
-                classes="new-expense-category small"
-              />
-            </Atom>
-            <AmountInput
-              name="amount"
-              label="Amount"
-              placeholder="0.00"
-              fieldError={fieldErrors.amount}
-              value={amount}
-              classes="big"
-              onFocus={() => setFieldErrors({ ...fieldErrors, amount: null })}
-              onChange={e => handleAmount(e.target.value)}
-            />
-          </div>
-          <div style={{
-            display: 'none',
-          }} className="split-details-page">
-            <>
-              {accountFetched && <AddUsers account={account}
-                label={'Split with'}
-                usersList={splitWith}
-                setUsersList={setSplitWith}
-                selectedGroup={splitWithGroup}
-                setOpenSelectPeople={setOpenSelectPeople}
-                removeGroupSelect={removeGroupSplit} />}
-              {fieldErrors.splitWith && (
-                <span className="field-error-text">{fieldErrors.splitWith}</span>
-              )}
-            </>
-            {splitWith.length > 1 && (
-              <>
-                <fieldset>
-                  <div className="expense-input">
-                    <legend className="input-label">Paid By</legend>
-                    <div className="expense-radio paid-by">
-                      {splitWith.map(member => {
-                        return (
-                          <UserButton
-                            key={member?.id + '-2'}
-                            user={member}
-                            name={member?.name}
-                            selected={paidBy === member?.id}
-                            onClick={() => {
-                              setPaidBy(member?.id);
-                              setFieldErrors({ ...fieldErrors, paidBy: null });
-                            }}
-                          />
-                        );
-                      })}
-                      {fieldErrors.paidBy && (
-                        <span className="field-error-text">
-                          {fieldErrors.paidBy}
-                        </span>
-                      )}
-                    </div>
-                  </div>
-                </fieldset>
-                <RadioSelect
-                  label="Split"
-                  name="expense-split"
-                  options={[
-                    {
-                      id: 'split-equally',
-                      value: 'EQUALLY',
-                      checked: split === 'EQUALLY',
-                      content: 'Split equally',
-                    },
-                    {
-                      id: 'full-amount',
-                      value: 'FULL_AMOUNT',
-                      checked: split === 'FULL_AMOUNT',
-                      content: 'Owed full amount',
-                    },
-                  ]}
-                  onFocus={() => setFieldErrors({ ...fieldErrors, split: null })}
-                  onChange={e => setSplit(e.target.value)}
-                  className="expense-input"
-                />
-                <button
-                  className="button"
-                  type="submit"
-                  alt="Create expense"
-                  title="Create expense"
-                >
-                  Create
-                </button>
-              </>
-            )}
-            {fieldErrors.formValid === false && (
-              <span className="field-error-text .form-validation">
-                Please fix the errors to submit expense.
-              </span>
-            )}
-          </div>
+          {page === 1 ? <NewExpenseDetails 
+            description={description}
+            setDescription={setDescription}
+            fieldErrors={fieldErrors}
+            setFieldErrors={setFieldErrors}
+            setPage={setPage}
+            amount={amount}
+            setAmount={setAmount}
+            handleValidation={handleValidation}
+          /> : null}
+          {page === 2 ? <NewExpenseSplit
+            account={account}
+            accountFetched={accountFetched}
+            paidBy={paidBy}
+            setPaidBy={setPaidBy}
+            splitWith={splitWith}
+            setSplitWith={setSplitWith}
+            splitWithGroup={splitWithGroup}
+            split={split}
+            setSplit={setSplit}
+            removeGroupSplit={removeGroupSplit}
+            setOpenSelectPeople={setOpenSelectPeople}
+            fieldErrors={fieldErrors}
+            setFieldErrors={setFieldErrors}
+          /> : null}
         </form>
       </div>
       <Modal open={openSelectPeople}>
