+++ src/pages/NewExpense/NewExpense.scss
@@ -8,27 +8,27 @@
 .expense-container {
   height: 100%;
   width: 100%;
+  .page-button {
+    position: absolute;
+    bottom: 1rem;
+    width: calc(100% - 40px);
+  }
 }
 
 .expense-form {
   height: 100%;
-  .expense-input {
-    .field-error {
-      border: 2px solid v.$red;
-    }
-  }
 
   .amount-input-container {
     z-index: 3;
-    .field-error-text {
-      display: none;
-    }
   }
 
   .expense-input-container {
     z-index: 3;
     display: grid;
     gap: 1.5rem;
+    .text-input-container {
+      align-items: center;
+    }
   }
 
   .expense-radio.paid-by {
@@ -50,7 +50,7 @@
   display: flex;
   flex-direction: column;
   margin-top: 15%;
-  gap: 60px;
+  gap: 70px;
 
   .atom--fade {
     &::before {
